debug: move debugContentProvider to common
