+++ src/vs/workbench/contrib/debug/common/debugContentProvider.ts
